Create user-extensions.js
Concatenates all the files from the extensions folder in the same order as is listed in extension-loader.xul.
Adds `testCase` as a global var. changes all `testCase.debugContext` to just `testCase`.
Fixes `nextCommand` so it will work on the server.
Fixes `setNextCommand` to work on the server.
Adds `mapList` for importing the collection of commands onto `testCase`.
Fixes `$$.fn.interceptAfter(Selenium.prototype, "reset", function` to set up `testCase` on the server.
Extends `HtmlRunnerTestLoop.prototype.initialize` to set up `testCase`.

I've tested this file on FireFox 34.0.5 a little bit. It appears to allow me to use functions and if statements in Selenese scripts. I haven't tried anything else, and what I did try was as simple as possible.