adds conditionally executed code for RC server
The new build script will generate user-extensions.js. The new directory
"user extension" is for the output user-extension.js and can be used for
scripts that may need to be included on the server only. selblocks.js
has been altered so that the selblocks object has a new property
"onServer" that is false by default. After checking for the existence of
"testCase" in the global context, onServer will be set to true if
"testCase" does not exist. Using that flag, additional code will be
executed if the server environment is detected.