sets whether on server or not by config.js
Detecting whether this is running on the server or not is silly. I just
include a setting when creating the user-extension. If the setting is
set, then extra code for the server gets executed. If the setting is
undefined, then the server code is not executed. The extension passes
the test suite and the server extension doesn't crash on the server. Now
I'll figure out SelBench and we'll be able to run the full test suite on
the server.