moved server augmentation code to config.js
This moves the server augmentation code out of the FF extension. This
appears to be working, I'm missing some mappings between obejcts and
methods on the server and IDE though. This passes testing in the IDE but
fails gloriously on the server. The "while" test makes the screen twitch
because the incrementor never increments. I'm sure it's something simple
once I look...