Moved server root for debug tests
When giving the server an html suite to run automatically, it will root
there. This overrides setting the server root by placement of the
user-extensions.js file. To make file paths consistent between automatic
and debug tests, I moved the location of the debug copy of the user
script into the sel-blocksTests folder.