added paths for testing file ops on the server
The IDE loads the test suite and sets the current working directory to
wherever you loaded the test suite from, I think. That might be a
function of the path manipulations done as well. I don't manipulate the
URL given to forJson and forXml, since data could exist anywhere on the
internet and the browser will resolve the url anyway.