issue fixed: functions aren't global
Since the blockdefs and stack are reset on a per testCase basis, the
symbols need to be reset as well. If they're not, then the function
names persist between test cases but the blockdefs don't. This causes
selblocks to execute the range of indexes stored for the function, but
using the commands array from the current case, which is obviously
wrong. Resetting the symbols makes the behavior "functions aren't
global", correct and fixes the bug.