Beginning specs for calling caseName.fn
calling functions from other test cases should require qualifying the
function name with the test case title, that way two cases could have a
function with the same name and not collide. Think of it like two
namespaces. Currently, this doesn't execute functions in other cases, it
just uses the cache to lookup the symbol for the current case.