moved saving vars into doFunction
we have to cache the values of all storedVars that exist in the function
signature, since we're initializing them to undefined for the function
block. If we don't cache all the storedVars spec'd in the sig., then
when the call completes only the args specified in the call will be
restored while the unspecified params will be left as undefined on the
storedVars object, or leak whatever value may have been assigned to them
in the function call.