storeLocal and storeGlobal work, funcs have blocks
I've implemented block scope for function variables. storeLocal stores
local to the block, storeGlobal stores to the global variables. Both
local and global variables can be expanded when using the ${} notation.
The original store commands, or anything that tries to write directly to
storedVars will be subject to the new block scope mechanism.

To start a new block scope just write "contextManager.enter()", to
return to the parent block scope just write "contextManager.exit()" They
should be able to nest for as long and large as a javascript prototype
chain is allowed to be.