update test for block scope
The iterator value "w" is evaluated and stored inside the "with" block.
The iterator emits the right values but after the block is completed,
emitting "end ${w}" gets the value of w in the scope above the "with"
block, which hasn't been changed.

The variable is automatically shadowed in local scope, so storing the
iterator value globally doesn't work because the updated value isn't
visible locally. I don't know how that works exactly, it's something
with the way arguments are being parsed.