Passes testing in all but IE
Everything is working as expected in the IDE, and on the server in
Firefox and Chrome. Internet Explorer passes all the tests except for
calling functions across cases. I don't know if it's because the
functions I'm testing are calling slow web pages or if it's some kind of
syntax error in IE when we use the command name "call". IE is cool and
all, but I'm submitting this because it works everywhere else and one of
us will be able to figure out the bug in IE in the near future.