Fixes doCall bug in IE
when doCall tries to validate the "argSpec" argument, the validation
method was throwing an error in IE if the value was a blank string. We
want to be able to call functions without any arguments, so I told it to
skip validation when the argSpec is like a blank string. Like a blank
string, not exactly a blank string, but like a blank string. parseArgs
handles this just fine.