Refactoring tests to test just selblocks
there's some kind of bug with selenium in IE, it tries to use the window
before it's ready and that is causing our tests to fail if they open
pages. I've changed the tests so we're not getting false failures
anymore, except for in the readme test, since all it does is open a
window.